Conflicts:
	lib/active_record/connection_adapters/oracle_enhanced_base_ext.rb
	lib/active_record/connection_adapters/oracle_enhanced_schema_statements.rb
	spec/active_record/connection_adapters/oracle_enhanced_schema_dump_spec.rb

Conflict resolutions

  #1 lib/active_record/connection_adapters/oracle_enhanced_base_ext.rb

  This file has been removed in rails4 branch at 9369d15b77da24c6fd06928d90cbe755a1f6982b
  "Initial support for ActiveRecord 4. Dropping support for older versions."

  Several changes done with 3fad716c0132af0e833b2da26738642500326408 are
  merged to `lib/active_record/connection_adapters/oracle_enhanced_model.rb`

  #2 lib/active_record/connection_adapters/oracle_enhanced_schema_statements.rb

  These two commits at master branch conflict with 166bcb21f7b617be1f0c62affc73bacc476dac4e

  -- Two commits at master branch
  305283a8e762e625f00789dbe22e45d25075b710
  "Remove_column should raise an ArgumentError when no columns are passed"

  38e7b728476fe8133be24f7a3cf6f4572c250811
  "Fix broken column remove inside a change_table block."

  -- A commit at rails4 branch
  166bcb21f7b617be1f0c62affc73bacc476dac4e
  "Support refactored remove_column"

  #3 spec/active_record/connection_adapters/oracle_enhanced_schema_dump_spec.rb

  This commit done at master branch conflict with rails4.
  e35a4a00ca2e9219807246f1dd9cd27e3943c4be
  "Fix pending specs on Oracle 10g properly"

+++ spec/active_record/connection_adapters/oracle_enhanced_schema_dump_spec.rb
@@ -347,33 +347,40 @@ describe "OracleEnhancedAdapter schema dump" do
 
   describe 'virtual columns' do
     before(:all) do
-      pending "Not supported in this database version" unless @oracle11g
-      schema_define do
-        create_table :test_names, force: true do |t|
-          t.string  :first_name
-          t.string  :last_name
-          t.virtual :full_name,        as: "first_name || ', ' || last_name"
-          t.virtual :short_name,       as: "COALESCE(first_name, last_name)", type: :string, limit: 300
-          t.virtual :abbrev_name,      as: "SUBSTR(first_name,1,50) || ' ' || SUBSTR(last_name,1,1) || '.'", type: "VARCHAR(100)"
-          t.column  :full_name_length, :virtual, as: "length(first_name || ', ' || last_name)", type: :integer
-          t.virtual :field_with_leading_space, as: "' ' || first_name || ' '", limit: 300, type: :string
+      if @oracle11g
+        schema_define do
+          create_table :test_names, :force => true do |t|
+            t.string  :first_name
+            t.string  :last_name
+            t.virtual :full_name,        :as => "first_name || ', ' || last_name"
+            t.virtual :short_name,       :as => "COALESCE(first_name, last_name)", :type => :string, :limit => 300
+            t.virtual :abbrev_name,      :as => "SUBSTR(first_name,1,50) || ' ' || SUBSTR(last_name,1,1) || '.'", :type => "VARCHAR(100)"
+            t.column  :full_name_length, :virtual, :as => "length(first_name || ', ' || last_name)", :type => :integer
+            t.virtual :field_with_leading_space, :as => "' ' || first_name || ' '", :limit => 300, :type => :string
+          end
         end
+      else
+        pending "Not supported in this database version"
       end
     end
 
     before(:each) do
-      class ::TestName < ActiveRecord::Base
-        if self.respond_to?(:table_name=)
-          self.table_name = "test_names"
-        else
-          set_table_name "test_names"
+      if @oracle11g
+        class ::TestName < ActiveRecord::Base
+          if self.respond_to?(:table_name=)
+            self.table_name = "test_names"
+          else
+            set_table_name "test_names"
+          end
         end
       end
     end
 
     after(:all) do
-      schema_define do
-        drop_table :test_names
+      if @oracle11g
+        schema_define do
+          drop_table :test_names
+        end
       end
     end
 
@@ -403,6 +410,27 @@ describe "OracleEnhancedAdapter schema dump" do
 
         standard_dump
         col.virtual_column_data_default.should_not =~ /:as/
+      end
+    end
+
+    context "with index on virtual column" do
+      before(:all) do
+        if @oracle11g
+          schema_define do 
+            add_index 'test_names', 'field_with_leading_space', :name => "index_on_virtual_col"
+          end
+        end
+      end
+      after(:all) do
+        if @oracle11g
+          schema_define do
+            remove_index 'test_names', :name => 'index_on_virtual_col'
+          end
+        end
+      end
+      it 'should dump correctly' do
+        standard_dump.should_not =~ /add_index "test_names".+FIRST_NAME.+$/
+        standard_dump.should     =~ /add_index "test_names".+field_with_leading_space.+$/
       end
     end
   end
